Cache photos in system's tmp directory
Use the temporary directory as a cache.

Name each image directory after the ID of the photo.

Put each image directory (still one photo per directory) directly in the
system temporary directory; if they ar bundeled in a subdirectory, that
subdirectory might never be disposed of because it keeps being modified.

Don't make an image directory if no valid image is returned from Photos.

Rationale

The tmp module is not deleting the temporary folders that it is supposed
to be manageing.

MacOs seems to do a pretty good job of removing oldthings from its
temporary directory that haven't been accessed in a while.